borrar archivos sin usar, arreglar bug que no dejaba que apareciera aplicacion centrada

+++ ui_functions.py
@@ -34,8 +34,8 @@ def centerOnScreen (self):
         '''centerOnScreen()
 Centers the window on the screen.'''
         resolution = QDesktopWidget().screenGeometry()
-        self.move((resolution.width() / 2) - (self.frameSize().width() / 2),
-                  (resolution.height() / 2) - (self.frameSize().height() / 2)) 
+        self.move(int((resolution.width() / 2) - (self.frameSize().width() / 2)),
+                  int((resolution.height() / 2) - (self.frameSize().height() / 2))) 
     # MAXIMIZE/RESTORE
     # ///////////////////////////////////////////////////////////////
 def maximize_restore(self):
